fastapi_template/requirements.txt
uvicorn==0.29
fastapi==0.111.0
fastapi-utils==0.6.0
fpdf2==2.7.8
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.1
requests==2.31.0
passlib[bcrypt]
tenacity==8.3.0
pydantic==2.7.1
emails==0.6
raven==6.10.0
gunicorn==22.0.0
jinja2==3.1.4
aiofiles==23.2.1
psycopg2-binary==2.9.9
importlib-metadata==7.1.0
alembic==1.13.1
sqlalchemy==2.0.30
pytest==8.2.0
websockets==12.0
watchfiles==0.21.0
pydantic-settings==2.2.1
python-jose[extras,cryptography]
pymysql==1.1.0
wheel
regex
pandas==2.0.3
sqlalchemy-stubs==0.4
pydantic[email]
pytest-cov==4.1.0
